Add regression test for processors
