+++ tests/logger_structlog/test_log_forwarding.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import pytest
+
 from newrelic.api.background_task import background_task
 from testing_support.fixtures import override_application_settings, reset_core_stats_engine
 from testing_support.validators.validate_log_event_count import validate_log_event_count
@@ -26,6 +28,27 @@ _common_attributes_service_linking = {"timestamp": None, "hostname": None,
 
 _common_attributes_trace_linking = {"span.id": "abcdefgh", "trace.id": "abcdefgh12345678",
                                     **_common_attributes_service_linking}
+
+
+@pytest.fixture(scope="function")
+def callsite_parameter_logger(structlog_caplog):
+    import structlog
+
+    structlog.configure(
+        processors=(
+            structlog.processors.CallsiteParameterAdder(
+                [
+                    structlog.processors.CallsiteParameter.FILENAME,
+                    structlog.processors.CallsiteParameter.FUNC_NAME,
+                ],
+            ),
+            structlog.processors.KeyValueRenderer(),
+        ),
+        logger_factory=lambda *args, **kwargs: structlog_caplog,
+    )
+
+    _callsite_logger = structlog.get_logger()
+    return _callsite_logger
 
 
 @reset_core_stats_engine()
@@ -86,3 +109,18 @@ def test_logging_filtering_outside_transaction(exercise_filtering_logging_multip
         exercise_filtering_logging_multiple_lines()
 
     test()
+
+
+@reset_core_stats_engine()
+@override_application_settings({"application_logging.local_decorating.enabled": False})
+@validate_log_events([
+    {"message": "event='Dog' filename='test_log_forwarding.py' func_name='test_callsite_processor'", "level": "INFO"},
+])
+@validate_log_event_count(1)
+@background_task()
+def test_callsite_processor(callsite_parameter_logger, structlog_caplog):
+    callsite_parameter_logger.msg("Dog")
+    assert "Dog" in structlog_caplog.caplog[0]
+    assert len(structlog_caplog.caplog) == 1
+    assert "filename='test_log_forwarding.py'" in structlog_caplog.caplog[0]
+    assert "func_name='test_callsite_processor'" in structlog_caplog.caplog[0]
